// src/types/car.ts
export interface CarData {
  Manufacturer: string;
  Model: string;
  Sales_in_thousands: number;
  __year_resale_value: number;
  Vehicle_type: string;
  Price_in_thousands: number;
  Engine_size: number;
  Horsepower: number;
  Wheelbase: number;
  Width: number;
  Length: number;
  Curb_weight: number;
  Fuel_capacity: number;
  Fuel_efficiency: number;
  Latest_Launch: string;
  Power_perf_factor: number;
}